+++ app/Controllers/EntriesController.php
@@ -213,14 +213,14 @@ class EntriesController extends Container
             $parent_entry_id = '';
         }
 
-        // Set new Entry ID
+        // Set new Entry ID using slugify or without it
         if ($this->registry->get('plugins.admin.settings.entries.slugify') == true) {
-            $id = $parent_entry_id . '/' . $this->slugify->slugify($data['id']);
+            $id = ltrim($parent_entry_id . '/' . $this->slugify->slugify($data['id']), '/');
         } else {
-            $id = $parent_entry_id . '/' . $data['id'];
+            $id = ltrim($parent_entry_id . '/' . $data['id'], '/');
         }
 
-        // Check if entry exists then try to create
+        // Check if entry exists then try to create it
         if (!$this->entries->has($id)) {
 
             // Check if we have fieldset for this entry
@@ -229,11 +229,6 @@ class EntriesController extends Container
                 // Get fieldset
                 $fieldset = $this->fieldsets->fetch($data['fieldset']);
 
-                // We need to check if template for current fieldset is exists
-                // if template is not exist then `default` template will be used!
-                //$template_path = PATH['themes'] . '/' . $this->registry->get('flextype.theme') . '/templates/' . $data['fieldset'] . '.html';
-                //$template = (Filesystem::has($template_path)) ? $data['fieldset'] : 'default';
-
                 // Init entry data
                 $data_from_post          = [];
                 $data_from_post_override = [];
@@ -241,10 +236,18 @@ class EntriesController extends Container
 
                 // Define data values based on POST data
                 $data_from_post['title']      = $data['title'];
-                //$data_from_post['template']   = $template;
                 $data_from_post['fieldset']   = $data['fieldset'];
                 $data_from_post['visibility'] = $data['visibility'];
                 $data_from_post['routable']   = isset($data['routable']) ? (bool) $data['routable'] : false;
+
+                // Themes/Templates support for Site Plugin
+                // We need to check if template for current fieldset is exists
+                // if template is not exist then `default` template will be used!
+                if ($this->registry->has('plugins.site')) {
+                    $template_path = PATH['project'] . '/themes/' . $this->registry->get('plugins.site.settings.theme') . '/templates/' . $data['fieldset'] . '.html';
+                    $template = (Filesystem::has($template_path)) ? $data['fieldset'] : 'default';
+                    $data_from_post['template']   = $template;
+                }
 
                 // Predefine data values based on fieldset default values
                 foreach ($fieldset['sections'] as $section_name => $section_body) {
@@ -269,7 +272,6 @@ class EntriesController extends Container
                         }
 
                         $data_from_post_override[$field] = $value;
-
                     }
                 }
 
@@ -299,7 +301,7 @@ class EntriesController extends Container
         }
 
         if (isset($data['create-and-edit'])) {
-            return $response->withRedirect($this->router->pathFor('admin.entries.edit') . '?id=' . $data['id'] . '&type=editor');
+            return $response->withRedirect($this->router->pathFor('admin.entries.edit') . '?id=' . $id . '&type=editor');
         } else {
             return $response->withRedirect($this->router->pathFor('admin.entries.index') . '?id=' . $parent_entry_id);
         }
@@ -711,7 +713,7 @@ class EntriesController extends Container
                         'menu_item' => 'entries',
                         'links' => [
                             'entries' => [
-                                'link' => $this->router->pathFor('admin.entries.index') . '?id=' . implode('/', array_slice(explode("/", $this->getEntryID($query)), 0, -1)),
+                                'link' => $this->router->pathFor('admin.entries.index'),
                                 'title' => __('admin_entries'),
 
                             ],